uart: implement REQ-UART-034, widen read_size to the ACF header's full 12-bit field
TC18 13.7.8.1 explicitly contemplates a read_size larger than
uart_rx_fifo_size as the third UART read-completion trigger, driving
a fragmented response. rcp_ep_uart_encode_read_request()'s read_size
parameter and rcp_ep_uart_decode_read_request()'s *out_read_size
were narrowed to uint8_t (0-255), on the file header's own now-
corrected reasoning that 255 bytes always fits a single AVTPDU so
the fragmentation mechanism was never actually needed. That
reasoning didn't survive TC18's own text: a conforming peer's
request in the 256..4095 range was simply inexpressible.

Both are now uint16_t, matching the ACF header's own 12-bit
read_size_or_segment_num field. Updated every call site:
src/adapt.c's RCP_ADAPT_OP_UART_READ, plus test call sites across
tests/test_ep_uart.c and tests/test_tc18_gaps_ep2.c. Rewrote the
pre-existing gap-pinning test to assert the fixed round-trip, added
a new test verifying a value above 255 round-trips exactly.

Mutation-tested (reintroducing the 8-bit truncation on decode):
caught cleanly.

65/65 both trees. cfusa check: 0 errors. cfusa trace: 100%/100%.

Issue #201.

+++ src/ep_uart.c
@@ -469,7 +469,8 @@ rcp_ep_uart_errc_t rcp_ep_uart_decode_write_response(const uint8_t *b, size_t le
 /* ── RX: read request/response ─────────────────────────────────────────────── */
 
 //cfusa:req REQ-UART-023
-rcp_bytes_t rcp_ep_uart_encode_read_request(rcp_byte_bus_id_t byte_bus_id, uint8_t read_size,
+//cfusa:req REQ-UART-034
+rcp_bytes_t rcp_ep_uart_encode_read_request(rcp_byte_bus_id_t byte_bus_id, uint16_t read_size,
                                              uint8_t transaction_num)
 {
     rcp_acf_byte_message_info_t hdr = {0};
@@ -485,9 +486,10 @@ rcp_bytes_t rcp_ep_uart_encode_read_request(rcp_byte_bus_id_t byte_bus_id, uint8
 
 //cfusa:req REQ-UART-024
 //cfusa:req REQ-UART-025
+//cfusa:req REQ-UART-034
 rcp_ep_uart_errc_t rcp_ep_uart_decode_read_request(const uint8_t *b, size_t len,
                                                     rcp_byte_bus_id_t expected_bus_id,
-                                                    uint8_t *out_read_size,
+                                                    uint16_t *out_read_size,
                                                     uint8_t *out_transaction_num)
 {
     rcp_acf_byte_message_info_t hdr;
@@ -511,14 +513,11 @@ rcp_ep_uart_errc_t rcp_ep_uart_decode_read_request(const uint8_t *b, size_t len,
      * the file header. */
     if (payload_len != 0) return RCP_EP_UART_ERR_UNKNOWN_CMD;
 
-    /* out_read_size is this endpoint's own pre-existing octet-wide type
-     * (unchanged by the acf.c header rework); hdr.read_size_or_segment_num
-     * is now the wire header's real 12-bit uint16_t field, so the
-     * narrowing needs an explicit cast for MSVC's /W4 -- same reasoning as
-     * discovery.c's identical cast. A UART read request larger than 255
-     * bytes truncates here, same as it silently did before this pass; not
-     * a new limitation introduced by it. */
-    *out_read_size       = (uint8_t)hdr.read_size_or_segment_num;
+    /* FIXED 2026-08-12 (issue #201, REQ-UART-034): out_read_size is now
+     * the wire header's own 12-bit-wide uint16_t type -- no narrowing
+     * cast needed (or possible data loss) any more; see the header's own
+     * doc comment. */
+    *out_read_size       = hdr.read_size_or_segment_num;
     *out_transaction_num = hdr.transaction_num;
     return RCP_EP_UART_OK;
 }
